% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotness-package.R
\docType{package}
\name{plotness-package}
\alias{plotness-package}
\title{plotness: Poissonness, Binomialness, and Negative Binomialness Plots}
\description{
Creates a diagnostic distribution plot or data frame. The per-type metameter statistics are adapted from the 'vcd' package's distplot() function (GPL-2); this package produces 'ggplot2' graphics and can return only the data needed for the plot so the user's preferred graphics package can be used.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/zaynesember/plotness}
  \item \url{https://zaynesember.github.io/plotness/}
  \item Report bugs at \url{https://github.com/zaynesember/plotness/issues}
}

}
\author{
\strong{Maintainer}: Zayne Sember \email{zaynesember@gmail.com}

Other contributors:
\itemize{
  \item David Meyer (author of vcd::distplot(), from which the per-type statistics are adapted) [contributor, copyright holder]
  \item Achim Zeileis (author of vcd::distplot()) [contributor, copyright holder]
  \item Kurt Hornik (author of vcd::distplot()) [contributor, copyright holder]
  \item Michael Friendly (author of vcd::distplot()) [contributor, copyright holder]
}

}
\keyword{internal}
