% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotness.R
\name{plotness}
\alias{plotness}
\title{Poisonness, Binomialness, and Negative Binomialness Plots}
\usage{
plotness(
  x,
  type = c("poisson", "binomial", "nbinomial"),
  size = NULL,
  lambda = NULL,
  plot = TRUE,
  title = NULL,
  legend = TRUE,
  label = TRUE,
  conf_int = TRUE,
  conf_level = 0.95,
  xlim = NULL,
  xlab = "Number of occurrences",
  ylab = "Distribution metameter"
)
}
\arguments{
\item{x}{either a vector of counts, a 1-way table of frequencies of counts or
a data frame or matrix with frequencies in the first column and the
corresponding counts in the second column.}

\item{type}{a character string indicating the distribution.}

\item{size}{the size argument for the binomial and negative binomial
distribution. If set to NULL and type is "binomial", then size is taken to
be the maximum count. If set to NULL and type is "nbinomial", then size is
estimated from the data.}

\item{lambda}{parameter of the poisson distribution. If type is "poisson"
and lambda is specified a leveled Poissonness plot is produced.}

\item{plot}{logical. When `TRUE` a ggplot is returned; otherwise a data frame
containing data necessary for a plot is returned.}

\item{title}{a character string to be used as the title of the plot. If set
to `NULL` a generic title is produced based on the distribution type.}

\item{legend}{logical. Should a legend be plotted?}

\item{label}{logical. Should fit data be printed on the plot?}

\item{conf_int}{logical. Should confidence intervals be plotted?}

\item{conf_level}{confidence level for confidence intervals.}

\item{xlim}{limits for the x-axis.}

\item{xlab}{a label for the x-axis.}

\item{ylab}{a label for the y-axis.}
}
\value{
If `plot`=`TRUE`, returns a ggplot object; otherwise returns a
 data frame containing the data necessary to produce a plot.
}
\description{
Creates a diagnostic distribution plot or dataframe. The function is adapted
from the vcd library's distplot function to produce ggplots and to allow the
option to return only the data needed for the plot so the user's preferred
graphics package can be used.
}
\examples{
plotness(rpois(15,10),type="poisson");
plotness(rbinom(15,10),type="binomial");
plotness(rnbinom(15,10),type="binomial");
plotness(rpois(15,10),type="poisson", plot=FALSE);
}
\author{
Zayne Sember <zsember@ucsd.edu>
}
